links Shared code  (#228)
* refactor(chat/smart links): add a hook to share go-to and action handlers.

* chore(lsp links): add a note about the piece of shared code.

* refactor(chat and smart links): move goto handler to it's own hook and add a dedicated smart link hook.

* fix: use thread mode to find out if it's a configuration chat.
