+++ src/components/SmartLink/SmartLink.tsx
@@ -1,62 +1,8 @@
 import { useCallback } from "react";
 import type { FC } from "react";
-import type {
-  LspChatMessage,
-  SmartLink as SmartLinkType,
-} from "../../services/refact";
-import {
-  type OpenFilePayload,
-  useAppDispatch,
-  useEventsBusForIDE,
-} from "../../hooks";
-import { formatMessagesForChat } from "../../features/Chat/Thread/utils";
-import { clearInformation } from "../../features/Errors/informationSlice";
-import { newIntegrationChat } from "../../features/Chat";
-import { push } from "../../features/Pages/pagesSlice";
+import type { SmartLink as SmartLinkType } from "../../services/refact";
 import { Button, DropdownMenu } from "@radix-ui/themes";
-import { AppDispatch } from "../../app/store";
-
-const handleGotoAction = (
-  sl_goto: string,
-  queryPathThenOpenFile: (file: OpenFilePayload) => Promise<void>,
-) => {
-  const [action, payload] = sl_goto.split(":");
-  switch (action.toLowerCase()) {
-    // TODO: could be possible to share it between Marc's implementation
-    case "editor":
-      void queryPathThenOpenFile({ file_name: payload });
-      break;
-    case "setting":
-      // Handling SETTING smartlink action
-      break;
-    default:
-      // For unexpected actions
-      break;
-  }
-};
-
-const handleChatAction = (
-  sl_chat: LspChatMessage[],
-  dispatch: AppDispatch,
-  integrationName: string,
-  integrationPath: string,
-  integrationProject: string,
-) => {
-  const messages = formatMessagesForChat(sl_chat);
-
-  dispatch(clearInformation());
-  dispatch(
-    newIntegrationChat({
-      integration: {
-        name: integrationName,
-        path: integrationPath,
-        project: integrationProject,
-      },
-      messages,
-    }),
-  );
-  dispatch(push({ name: "chat" }));
-};
+import { useSmartLinks } from "../../hooks";
 
 export const SmartLink: FC<{
   smartlink: SmartLinkType;
@@ -73,21 +19,18 @@ export const SmartLink: FC<{
   isSmall = false,
   isDockerSmartlink = false,
 }) => {
-  const dispatch = useAppDispatch();
-
-  const { queryPathThenOpenFile } = useEventsBusForIDE();
+  const { handleGoTo, handleSmartLink } = useSmartLinks();
 
   const { sl_goto, sl_chat } = smartlink;
 
   const handleClick = useCallback(() => {
     if (sl_goto) {
-      handleGotoAction(sl_goto, queryPathThenOpenFile);
+      handleGoTo(sl_goto);
       return;
     }
     if (sl_chat) {
-      handleChatAction(
+      handleSmartLink(
         sl_chat,
-        dispatch,
         integrationName,
         integrationPath,
         integrationProject,
@@ -96,10 +39,10 @@ export const SmartLink: FC<{
   }, [
     sl_goto,
     sl_chat,
-    dispatch,
+    handleGoTo,
+    handleSmartLink,
     integrationName,
     integrationPath,
-    queryPathThenOpenFile,
     integrationProject,
   ]);
