+++ src/hooks/index.ts
@@ -22,3 +22,6 @@ export * from "./useAppSelector";
 export * from "./useSendChatRequest";
 export * from "./usePatchActions";
 export * from "./useGetUserSurvey";
+export * from "./useLinksFromLsp";
+export * from "./useGoToLink";
+export * from "./useSmartLinks";
