+++ src/utils/isAbsolutePath.ts
@@ -1,0 +1,4 @@
+const absolutePathRegex = /^(?:[a-zA-Z]:\\|\/|\\\\|\/\/).*/;
+export function isAbsolutePath(path: string): boolean {
+  return absolutePathRegex.test(path);
+}
